# gateway/gateway-controller/api/openapi.yaml
openapi: 3.0.3
info:
  title: Gateway Controller API
  description: |
    REST API for managing API configurations in the WSO2 API Platform Gateway.

    The Gateway Controller accepts API configurations in YAML or JSON format and dynamically
    configures the Router (Envoy Proxy) via xDS protocol. Supports full CRUD lifecycle
    for API configurations with real-time updates to the Router.

    **Key Features:**
    - Submit API configurations in YAML or JSON
    - Real-time validation and error reporting
    - Zero-downtime configuration updates
    - Persistent storage with SQLite
    - Query deployed API configurations
  version: 1.0.0
  contact:
    name: WSO2 API Platform Team
servers:
  - url: http://localhost:9090
    description: Local development
  - url: http://gateway-controller:9090
    description: Docker/Kubernetes deployment
paths:
...
  /apis:
    post:
      summary: Create a new API configuration
      description: |
        Submit a new API configuration to the Gateway Controller. The configuration
        is validated, persisted, and translated to Envoy xDS resources. The Router
        begins routing traffic for the API within 5 seconds.

        **Validation Rules:**
        - API name + version combination must be unique
        - Context must start with `/` and not end with `/`
        - At least one upstream and one operation required
        - For same API name, context must be consistent across versions
      operationId: createAPI
      tags:
        - API Management
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              $ref: "#/components/schemas/APIConfiguration"
          application/json:
            schema:
              $ref: "#/components/schemas/APIConfiguration"
            example:
              version: gateway.api-platform.wso2.com/v1alpha1
              kind: RestApi
              spec:
                name: PetStore API
                version: v1.0
                context: /petstore
                upstreams:
                  - url: http://petstore.swagger.io/v2
                policies:
                  - name: apiKeyValidation
                    version: v1
                    executionCondition: "request.metadata[authenticated] != true"
                    params:
                      header: X-API-Key
                      mandatory: true
                  - name: setHeaders
                    version: v1
                    params:
                      requestHeaders:
                        - X-Request-Source: WSO2 API Manager
                      responseHeaders:
                        - X-Response-Source: WSO2 API Manager
                operations:
                  - method: GET
                    path: /pet/{petId}
                    policies:
                      - name: apiKeyValidation
                        version: v1
                        executionCondition: "request.metadata[authenticated] != true"
                        params:
                          header: X-API-Key
                          mandatory: true
                  - method: POST
                    path: /pet
                  - method: PUT
                    path: /pet
                  - method: DELETE
                    path: /pet/{petId}
                  - method: GET
                    path: /store/inventory
                  - method: POST
                    path: /store/order
      responses:
        "201":
          description: API configuration created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APICreateResponse"
        "400":
          description: Invalid configuration (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict - API with same name and version already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      summary: List all API configurations
      description: |
        Retrieve a list of all deployed API configurations with metadata.
        Returns basic information about each API (name, version, context, deployment time).
      operationId: listAPIs
      tags:
        - API Management
      parameters:
        - name: displayName
          in: query
          required: false
          description: Filter by API display name
          schema:
            type: string
          example: Weather API
        - name: version
          in: query
          required: false
          description: Filter by API version
          schema:
            type: string
          example: v1.0
        - name: context
          in: query
          required: false
          description: Filter by API context/path
          schema:
            type: string
          example: /weather
        - name: status
          in: query
          required: false
          description: Filter by deployment status
          schema:
            type: string
            enum: [ pending, deployed, failed ]
          example: pending
      responses:
        "200":
          description: List of API configurations
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  count:
                    type: integer
                    example: 5
                  apis:
                    type: array
                    items:
                      $ref: "#/components/schemas/APIListItem"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /apis/{id}:
    get:
      summary: Get API configuration by id
      description: |
        Retrieve the complete configuration for a specific API by its id.
        The id is a unique string identifier (not a UUID).
      operationId: getAPIById
      tags:
        - API Management
      parameters:
        - name: id
          in: path
          required: true
          description: |
            Unique public identifier for the API.
          schema:
            type: string
          example: weather-api-v1.0
      responses:
        "200":
          description: API configuration details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIDetailResponse"
            application/yaml:
              schema:
                $ref: "#/components/schemas/APIDetailResponse"
        "404":
          description: API configuration not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      summary: Update an existing API configuration
      description: |
        Update an existing API configuration. The update is validated and applied atomically.
        The Router receives the updated configuration within 5 seconds. In-flight requests
        complete with the old configuration; new requests use the updated configuration.

        **Important:** If validation fails, the existing configuration remains active.
      operationId: updateAPI
      tags:
        - API Management
      parameters:
        - name: id
          in: path
          required: true
          description: |
            Unique public identifier of the API to update.
          schema:
            type: string
          example: weather-api-v1.0
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              $ref: "#/components/schemas/APIConfiguration"
          application/json:
            schema:
              $ref: "#/components/schemas/APIConfiguration"
      responses:
        "200":
          description: API configuration updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIUpdateResponse"
        "400":
          description: Invalid configuration (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: API configuration not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Delete an API configuration
      description: |
        Remove an API configuration from the Gateway Controller. The Router stops
        accepting traffic for this API within 5 seconds. In-flight requests complete
        successfully; new requests receive 404 Not Found.
      operationId: deleteAPI
      tags:
        - API Management
      parameters:
        - name: id
          in: path
          required: true
          description: |
            Unique public identifier of the API to delete.
          schema:
            type: string
          example: weather-api-v1.0
      responses:
        "200":
          description: API configuration deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: API configuration deleted successfully
                  id:
                    type: string
                    example: weather-api-v1.0
        "404":
          description: API configuration not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /apis/{id}/api-keys:
    post:
      summary: Create a new API key for an API
      description: |
        Generates a new API key for the specified API.
        Use the created key to authenticate client requests when the API Key validation policy is enabled.
        The generated key is a 32-byte random value, encoded in hexadecimal, and prefixed with apip_.
        If you are using an external API key, you may provide it directly without modification.
      operationId: createAPIKey
      tags:
        - API Management
      parameters:
        - name: id
          in: path
          required: true
          description: |
            Unique public identifier of the API to generate the key for
          schema:
            type: string
          example: weather-api-v1.0
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              $ref: "#/components/schemas/APIKeyCreationRequest"
          application/json:
            schema:
              $ref: "#/components/schemas/APIKeyCreationRequest"
      responses:
        '201':
          description: API key created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyCreationResponse"
        "400":
          description: Invalid configuration (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: API configuration not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      summary: Get the list of API keys for an API
      description: |
        Retrieve all active API keys for the specified API created by the user. If the user is an admin, retrieves all
        active API keys for the API.
      operationId: listAPIKeys
      tags:
        - API Management
      parameters:
        - name: id
          in: path
          required: true
          description: |
            Unique public identifier of the API to retrieve the keys for
          schema:
            type: string
          example: weather-api-v1.0
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyListResponse"
        "404":
          description: API configuration not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /apis/{id}/api-keys/{apiKeyName}:
    put:
      summary: Update an API key with a new regenerated value
      description: |
        Update an existing API key with a new regenerated value. This allows
        rotating API keys with a predetermined value instead of auto-generating one.
        The new key must be at least 20 characters long.
      operationId: updateAPIKey
      tags:
        - API Management
      parameters:
        - name: id
          in: path
          required: true
          description: |
            Unique public identifier of the API
          schema:
            type: string
          example: weather-api-v1.0
        - name: apiKeyName
          in: path
          required: true
          description: |
            Name of the API key to update
          schema:
            type: string
          example: weather-api-key
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              $ref: "#/components/schemas/APIKeyUpdateRequest"
          application/json:
            schema:
              $ref: "#/components/schemas/APIKeyUpdateRequest"
      responses:
        '200':
          description: API key updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyCreationResponse"
        "400":
          description: Invalid request (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: API or API key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Revoke an API key
      description: |
        Revoke a previously generated API key for the specified API. Once revoked,
        the key can no longer be used to authenticate requests.
      operationId: revokeAPIKey
      tags:
        - API Management
      parameters:
        - name: id
          in: path
          required: true
          description: |
            Unique public identifier of the API to revoke the key for
          schema:
            type: string
          example: weather-api-v1.0
        - name: apiKeyName
          in: path
          required: true
          description: |
            Name of the API key to revoke
          schema:
            type: string
          example: weather-api-key
      responses:
        '200':
          description: API key revoked successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyRevocationResponse"
        "400":
          description: Invalid configuration (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: API configuration not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    APIConfiguration:
      type: object
      required:
        - apiVersion
        - metadata
        - kind
        - spec
      properties:
        apiVersion:
          type: string
          description: API specification version
          example: gateway.api-platform.wso2.com/v1alpha1
          enum:
            - gateway.api-platform.wso2.com/v1alpha1
        kind:
          type: string
          description: API type
          example: RestApi
          enum:
            - RestApi
            - WebSubApi
        metadata:
          $ref: "#/components/schemas/Metadata"
        spec:
          description: API configuration payload (REST or Async API variants)
          anyOf:
            - $ref: '#/components/schemas/APIConfigData'
            - $ref: '#/components/schemas/WebhookAPIData'
    Metadata:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Unique handle for the resource
          example: weather-api-v1.0
        labels:
          type: object
          description: Labels are key-value pairs for organizing and selecting APIs. Keys must not contain spaces.
          additionalProperties:
            type: string
          example:
            environment: production
            team: backend
            version: v1
...
    APICreateResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: API configuration created successfully
        id:
          type: string
          description: Unique id for the created API configuration
          example: weather-api-v1.0
        created_at:
          type: string
          format: date-time
          example: 2025-10-11T10:30:00Z
    APIUpdateResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: API configuration updated successfully
        id:
          type: string
          example: weather-api-v1.0
        updated_at:
          type: string
          format: date-time
          example: 2025-10-11T11:45:00Z
    APIKeyCreationRequest:
      type: object
      properties:
        displayName:
          type: string
          description: Human-readable name for the API key (1-100 characters)
          minLength: 1
          maxLength: 100
          example: My Production Key
        name:
          type: string
          description: Identifier of the API key. If not provided, a default identifier will be generated
          pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
          minLength: 3
          maxLength: 63
          example: my-production-key
        apiKey:
          type: string
          minLength: 36
          description: |
            Optional plain-text API key value for external key injection.
            If provided, this key will be used instead of generating a new one.
            The key will be hashed before storage. The key can be in any format
            (minimum 36 characters). Use this for injecting externally generated
            API keys.
          example: "cloud-apim-key-abc123xyz789"
        expiresIn:
          type: object
          description: Expiration duration for the API key
          properties:
            unit:
              type: string
              description: Time unit for expiration
              enum:
                - seconds
                - minutes
                - hours
                - days
                - weeks
                - months
              example: days
            duration:
              type: integer
              description: Duration value for expiration
              example: 30
          required:
            - unit
            - duration
        expiresAt:
          type: string
          format: date-time
          description: Expiration timestamp. If both expiresIn and expiresAt are provided, expiresAt takes precedence.
          example: "2026-12-08T10:30:00Z"
        externalRefId:
          type: string
          description: |
            External reference ID for the API key.
            This field is optional and used for tracing purposes only.
            The gateway generates its own internal ID for tracking.
          example: "cloud-apim-key-98765"
    APIKeyCreationResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: API key generated successfully
        remainingApiKeyQuota:
          type: integer
          description: Remaining API key quota for the user
          example: 9
        apiKey:
          $ref: '#/components/schemas/APIKey'
      required:
        - status
        - message
...
    APIKeyUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/APIKeyCreationRequest"
      required:
        - apiKey
    APIKeyRevocationResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: API key revoked successfully
      required:
        - status
        - message
    APIListItem:
      type: object
      properties:
        id:
          type: string
          example: weather-api-v1.0
        displayName:
          type: string
          example: weather-api
        version:
          type: string
          example: v1.0
        context:
          type: string
          example: /weather
        status:
          type: string
          enum:
            - pending
            - deployed
            - failed
            - undeployed
          example: deployed
        created_at:
          type: string
          format: date-time
          example: 2025-10-11T10:30:00Z
        updated_at:
          type: string
          format: date-time
          example: 2025-10-11T10:30:00Z
    APIDetailResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        api:
          type: object
          properties:
            id:
              type: string
              description: Unique id for the API configuration
              example: weather-api-v1.0
            configuration:
              $ref: "#/components/schemas/APIConfiguration"
            metadata:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - pending
                    - deployed
                    - failed
                  example: deployed
                created_at:
                  type: string
                  format: date-time
                  example: 2025-10-11T10:30:00Z
                updated_at:
                  type: string
                  format: date-time
                  example: 2025-10-11T10:30:00Z
                deployed_at:
                  type: string
                  format: date-time
                  example: 2025-10-11T10:30:05Z
...
    ErrorResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          description: High-level error description
          example: Configuration validation failed
        errors:
          type: array
          description: Detailed validation errors
          items:
            $ref: "#/components/schemas/ValidationError"
    ValidationError:
      type: object
      properties:
        field:
          type: string
          description: Field that failed validation
          example: spec.context
        message:
          type: string
          description: Human-readable error message
          example: Context must start with / and cannot end with /
...
    APIKeyListResponse:
      type: object
      properties:
        apiKeys:
          type: array
          items:
            $ref: "#/components/schemas/APIKey"
        totalCount:
          type: integer
          description: Total number of API keys
          example: 3
        status:
          type: string
          example: success
...